requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.30
pandas-ta>=0.3.14b

# Database & API
supabase>=2.0.0
postgrest>=0.10.0
realtime>=1.0.0
gotrue>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
pytz>=2023.3
colorama>=0.4.6

# HTML Parsing (Required for pd.read_html ticker retrieval)
lxml>=4.9.0
html5lib>=1.1
beautifulsoup4>=4.12.0

# Optional/Transitive fixes
urllib3>=2.0.0
certifi>=2024.0.0
